requirements.txt
pillow==10.4.0
pytesseract==0.3.10
pygame==2.6.1
keyboard==0.13.5
deep-translator==1.11.4
edge-tts==6.1.3
ttkbootstrap==1.10.1
